+++ Changes.md
@@ -14,6 +14,7 @@ Improvements
 - SceneInspector : Added Interpretation row for primitive variables, with values of "Point", "Vector", "Normal" or "UV".
 - ImageToTensor : Added `tensorElementType` plug to choose between `Float`, `Float16` and `BFloat16`.
 - TensorToImage : Added conversion from `Float16` and `BFloat16` tensor elements in addition to the existing `Float` support.
+- RenderManOptions : Added XPU device configuration options.
 
 Fixes
 -----